+++ lib/onlyoffice_digitalocean_wrapper/digitalocean_wrapper/digitalocean_exceptions.rb
@@ -2,4 +2,8 @@ module OnlyofficeDigitaloceanWrapper
   # Class raised if image with name is not found
   class DigitalOceanImageNotFound < StandardError
   end
+
+  # Class raised if droplet size is not supported
+  class DigitalOceanSizeNotSupported < StandardError
+  end
 end