+++ Changelog.md
@@ -3,6 +3,7 @@
 # unreleased (master)
 * Add check of droplet size for creation
 * Do not force 2.0.1 of `droplet_kit`
+* Change default region to `nyc3`
 
 # 0.2.0
 * Force use ruby 2.2 to be able use `activesupport 5`