+++ Changelog.md
@@ -13,6 +13,7 @@
 
 ### changes
 * `wait_until_droplet_have_status` now raise DropletOperationTimeout if timeout reached
+* Remove `activesupport` dependency
 
 # 0.2.0
 * Force use ruby 2.2 to be able use `activesupport 5`